requirements.txt
django 
channels 
channels_redis 
openai 
langchain
fastapi
uvicorn
langchain-anthropic
python-dotenv
websockets